+++ docs/website/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.35.29](https://github.com/ttoss/ttoss/compare/@docs/website@1.35.28...@docs/website@1.35.29) (2023-04-26)
+
+**Note:** Version bump only for package @docs/website
+
+## [1.35.28](https://github.com/ttoss/ttoss/compare/@docs/website@1.35.27...@docs/website@1.35.28) (2023-04-25)
+
+**Note:** Version bump only for package @docs/website
+
+## [1.35.27](https://github.com/ttoss/ttoss/compare/@docs/website@1.35.26...@docs/website@1.35.27) (2023-04-25)
+
+**Note:** Version bump only for package @docs/website
+
+## [1.35.26](https://github.com/ttoss/ttoss/compare/@docs/website@1.35.25...@docs/website@1.35.26) (2023-04-25)
+
+**Note:** Version bump only for package @docs/website
+
+## [1.35.25](https://github.com/ttoss/ttoss/compare/@docs/website@1.35.24...@docs/website@1.35.25) (2023-04-25)
+
+**Note:** Version bump only for package @docs/website
+
+## [1.35.24](https://github.com/ttoss/ttoss/compare/@docs/website@1.35.23...@docs/website@1.35.24) (2023-04-25)
+
+**Note:** Version bump only for package @docs/website
+
 ## [1.35.23](https://github.com/ttoss/ttoss/compare/@docs/website@1.35.22...@docs/website@1.35.23) (2023-04-24)
 
 **Note:** Version bump only for package @docs/website