+++ packages/components/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.26.42](https://github.com/ttoss/ttoss/compare/@ttoss/components@1.26.41...@ttoss/components@1.26.42) (2023-04-26)
+
+**Note:** Version bump only for package @ttoss/components
+
+## [1.26.41](https://github.com/ttoss/ttoss/compare/@ttoss/components@1.26.40...@ttoss/components@1.26.41) (2023-04-25)
+
+**Note:** Version bump only for package @ttoss/components
+
 ## [1.26.40](https://github.com/ttoss/ttoss/compare/@ttoss/components@1.26.39...@ttoss/components@1.26.40) (2023-04-24)
 
 **Note:** Version bump only for package @ttoss/components